+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog
 =========
+
+## Unreleased
+### Changed
+- Disabled attachment/document downloads by default.
+
 ## 1.3.0 - 2019-9-30
 ### Added
 - Attachment download capabilities